change parameter type of coordinate (int -> double)

+++ BrickInvaders/src/main/java/org/example/Interfaces/Movable.java
@@ -2,5 +2,5 @@ package org.example.Interfaces;
 
 public interface Movable {
     void show();
-    void move(int xCoordinate, int yCoordinate);
+    void move(double xCoordinate, double yCoordinate);
 }
